requirements-minimal.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
mysql-connector-python==9.4.0
pymysql==1.1.1
python-dotenv
